+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Added
 
 - _Constant expression_ inspection: check expression overflow, number value, instruction operand value
+- Editor Color Scheme settings: highlight `rem`/`erem` block as _Deleted Text_ (_Comments | Rem/Erem block_)
 
 ### Fixed